work on train supernet

+++ model/gate_model.py
@@ -167,7 +167,7 @@ class AVnet_Gate(nn.Module):
 
         compress = [(torch.argmax(gate_a, dim=-1).float().mean() + 1).item()/12 ,
               (torch.argmax(gate_i, dim=-1).float().mean() + 1).item()/12,
-            (torch.argmax(gate_a, dim=-1) - torch.argmax(gate_i, dim=-1)).abs().mean().item()/12]
+            (torch.argmax(gate_a, dim=-1) - torch.argmax(gate_i, dim=-1)).float().abs().mean().item()/12]
         acc = (torch.argmax(output, dim=-1) == label).sum().item() / len(label)
 
         # print(loss_c.item(), (loss_g1 + loss_g2).item(), loss_r.item())
